Add docker support

+++ config.nims
@@ -6,15 +6,17 @@ mode = ScriptMode.Verbose
 let
   binaryName    = "luntic"
   mainFileName  = "main"
-  buildFolder               = "build"
+  version       = "1.1.0"
+  buildFolder               = thisDir() & '/' & "build"
   buildSourcesMainFolder    = buildFolder & "/sources/main"
   buildSourcesTestFolder    = buildFolder & "/sources/test"
   buildCacheFolder          = buildFolder & "/cache"
   buildTargetFolder         = buildFolder & "/target"
+  buildDockerFolder         = buildFolder & "/docker"
   buildResourcesMainFolder  = buildFolder & "/resources/main"
   buildResourcesTestFolder  = buildFolder & "/resources/test"
   buildBinaryFile           = buildTargetFolder & "/" & binaryName
-  sourcesFolder               = "src"
+  sourcesFolder               = thisDir() & '/' & "src"
   sourcesMainFolder           = sourcesFolder & "/main"
   sourcesMainNimFolder        = sourcesMainFolder & "/nim"
   sourcesMainResourcesFolder  = sourcesMainFolder & "/resources"
@@ -42,17 +44,17 @@ proc build_copy () =
   build_create()
   if dirExists sourcesMainNimFolder:
     exec "cp -r $1/* $2".format(sourcesMainNimFolder, buildSourcesMainFolder)
-  if dirExists sourcesMainResourcesFolder:
+  if dirExists(sourcesMainResourcesFolder) and listFiles(sourcesMainResourcesFolder).len > 0:
     exec "cp -r $1/* $2".format(sourcesMainResourcesFolder, buildResourcesMainFolder)
 
 
 proc test_copy () =
   build_copy()
-  if dirExists sourcesTestNimFolder:
+  if dirExists(sourcesTestNimFolder) and listFiles(sourcesTestNimFolder).len > 0:
     if not dirExists buildSourcesTestFolder:
       mkdir buildSourcesTestFolder
     exec "cp -r $1/* $2".format(sourcesTestNimFolder, buildSourcesTestFolder)
-  if dirExists sourcesTestResourcesFolder:
+  if dirExists(sourcesTestResourcesFolder) and listFiles(sourcesTestResourcesFolder).len > 0:
     if not dirExists buildResourcesTestFolder:
       mkdir buildResourcesTestFolder
     exec "cp -r $1/* $2".format(sourcesTestResourcesFolder, buildResourcesTestFolder)
@@ -134,7 +136,29 @@ task build, "builds the project":
 task init, "initialize a project":
   for folder in @[sourcesMainNimFolder, sourcesMainResourcesFolder, sourcesTestNimFolder, sourcesTestResourcesFolder]:
     mkdir folder
-  exec "echo 'echo \"Hello world\"' > $1".format(sourcesMainFile)
+  exec "echo 'echo \"Hello world\"' > $1/$2.nim".format(sourcesMainNimFolder, mainFileName)
+
+
+task docker, "builds the project within specific docker container":
+  dependsOn clean
+  build_copy()
+
+  exec "mkdir -p $1".format(buildDockerFolder)
+
+  exec """docker run \
+--rm \
+-v '$1:/tmp/src' \
+-v '$2:/tmp/target' \
+xxlabaza/nim:0.17.0 \
+compile --define:release \
+        --symbolFiles:off \
+        --nimcache:/tmp/src/cache \
+        --path:/tmp/src/sources/main \
+        --path:/tmp/src/resources/main \
+        --out:/tmp/target/$3 \
+        /tmp/src/sources/main/$4""".format(buildFolder, buildDockerFolder, binaryName, mainFileName)
+
+  exec "docker build --force-rm=true --tag xxlabaza/luntic:$1 .".format(version)
 
 
 task run, "runs the project":
